+++ spec/schleuder/integration/keywords_spec.rb
@@ -16,7 +16,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-SUBSCRIBE: test@example.org\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-SUBSCRIBE: test@example.org"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -59,7 +59,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-SUBSCRIBE: test@example.org 0x#{list.fingerprint} true false\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-SUBSCRIBE: test@example.org 0x#{list.fingerprint} true false"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -101,7 +101,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-SUBSCRIBE: test@example.org 0x#{list.fingerprint.dup.insert(4, ' ')} true\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-SUBSCRIBE: test@example.org 0x#{list.fingerprint.dup.insert(4, ' ')} true"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -144,7 +144,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-SUBSCRIBE: test@example.org 0x#{list.fingerprint.dup.insert(4, ' ')}\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-SUBSCRIBE: test@example.org 0x#{list.fingerprint.dup.insert(4, ' ')}"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -186,7 +186,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-SUBSCRIBE: test@example.org 0x#{list.fingerprint.dup.insert(4, ' ')} true false\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-SUBSCRIBE: test@example.org 0x#{list.fingerprint.dup.insert(4, ' ')} true false"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -228,7 +228,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-SUBSCRIBE: test@example.org 0x#{list.fingerprint.dup.insert(4, ' ')} false false\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-SUBSCRIBE: test@example.org 0x#{list.fingerprint.dup.insert(4, ' ')} false false"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -270,7 +270,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-SUBSCRIBE: test@example.org 0x#{list.fingerprint.dup.insert(4, ' ')} false true\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-SUBSCRIBE: test@example.org 0x#{list.fingerprint.dup.insert(4, ' ')} false true"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -312,7 +312,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-SUBSCRIBE:\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-SUBSCRIBE:"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -351,7 +351,7 @@ describe 'user sends keyword' do
       sign_as: '59C71FB38AEE22E091C78259D06350440F759BD3'
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-UNSUBSCRIBE:\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-UNSUBSCRIBE:"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -387,7 +387,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-UNSUBSCRIBE: test@example.org\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-UNSUBSCRIBE: test@example.org"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -421,7 +421,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-UNSUBSCRIBE: test@example.org\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-UNSUBSCRIBE: test@example.org"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -458,7 +458,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-UNSUBSCRIBE: schleuder@example.org\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-UNSUBSCRIBE: schleuder@example.org"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -493,7 +493,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-set-fingerprint: schleuder@example.org C4D60F8833789C7CAA44496FD3FFA6613AB10ECE\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-set-fingerprint: schleuder@example.org C4D60F8833789C7CAA44496FD3FFA6613AB10ECE"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -532,7 +532,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-set-fingerprint: schleuder@example.org C4D6 0F88 3378 9C7C  AA44 496F D3FF A661 3AB1 0ECE\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-set-fingerprint: schleuder@example.org C4D6 0F88 3378 9C7C  AA44 496F D3FF A661 3AB1 0ECE"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -571,7 +571,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-set-fingerprint: C4D60F8833789C7CAA44496FD3FFA6613AB10ECE\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-set-fingerprint: C4D60F8833789C7CAA44496FD3FFA6613AB10ECE"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -610,7 +610,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-set-fingerprint: test@example.org C4D60F8833789C7CAA44496FD3FFA6613AB10ECE\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-set-fingerprint: test@example.org C4D60F8833789C7CAA44496FD3FFA6613AB10ECE"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -649,7 +649,7 @@ describe 'user sends keyword' do
       sign_as: '59C71FB38AEE22E091C78259D06350440F759BD3'
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-set-fingerprint: test@example.org 59C71FB38AEE22E091C78259D06350440F759BD3\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-set-fingerprint: test@example.org 59C71FB38AEE22E091C78259D06350440F759BD3"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -687,7 +687,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-set-fingerprint: schleuder@example.org \nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-set-fingerprint: schleuder@example.org "
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -728,7 +728,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-set-fingerprint: schleuder@example.org 59C71FB38AEE22E091C78259D0\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-set-fingerprint: schleuder@example.org 59C71FB38AEE22E091C78259D0"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -768,7 +768,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-set-fingerprint: blabla\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-set-fingerprint: blabla"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -807,7 +807,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-set-fingerprint: bla@example.org C4D60F8833789C7CAA44496FD3FFA6613AB10ECE\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-set-fingerprint: bla@example.org C4D60F8833789C7CAA44496FD3FFA6613AB10ECE"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -840,7 +840,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-set-fingerprint: \nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-set-fingerprint: "
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -873,7 +873,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-unset-fingerprint: \nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-unset-fingerprint: "
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -908,7 +908,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-unset-fingerprint: test@example.org\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-unset-fingerprint: test@example.org"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -945,7 +945,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-unset-fingerprint: schleuder@example.org\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-unset-fingerprint: schleuder@example.org"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -982,7 +982,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-unset-fingerprint: schleuder@example.org force\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-unset-fingerprint: schleuder@example.org force"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1018,7 +1018,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-unset-fingerprint: bla@example.org\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-unset-fingerprint: bla@example.org"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1053,7 +1053,7 @@ describe 'user sends keyword' do
       sign_as: '59C71FB38AEE22E091C78259D06350440F759BD3'
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-unset-fingerprint: test@example.org\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-unset-fingerprint: test@example.org"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1091,7 +1091,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-list-subscriptions\nx-stop:"
+    mail.body = "x-list-name: #{list.email}\nX-list-subscriptions:"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1126,7 +1126,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-list-subscriptions:\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-list-subscriptions:"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1167,7 +1167,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-list-subscriptions: example.org\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-list-subscriptions: example.org"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1201,7 +1201,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-list-subscriptions: blabla\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-list-subscriptions: blabla"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1236,7 +1236,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-fetch-KEY: lala!\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-fetch-KEY: lala!"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1270,7 +1270,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-fetch-KEY: admin@example.org\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-fetch-KEY: admin@example.org"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1306,7 +1306,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-fetch-KEY: something@localhost\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-fetch-KEY: something@localhost"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1342,7 +1342,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-fetch-KEY: http://127.0.0.1:9999/keys/example.asc\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-fetch-KEY: http://localhost:9999/keys/example.asc"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1379,7 +1379,7 @@ describe 'user sends keyword' do
     }
     mail.gpg(gpg_opts)
     url = 'http://localhost:9999/foo'
-    mail.body = "x-list-name: #{list.email}\nX-fetch-KEY: #{url}\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-fetch-KEY: #{url}"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1415,7 +1415,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-fetch-KEY: 0x0000000000000000000000000000000000000000\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-fetch-KEY: 0x0000000000000000000000000000000000000000"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1451,7 +1451,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-fetch-KEY: 0x98769E8A1091F36BD88403ECF71A3F8412D83889\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-fetch-KEY: 0x98769E8A1091F36BD88403ECF71A3F8412D83889"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1487,7 +1487,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-fetch-KEY: 0x59C71FB38AEE22E091C78259D06350440F759BD3\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-fetch-KEY: 0x59C71FB38AEE22E091C78259D06350440F759BD3"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1523,7 +1523,7 @@ describe 'user sends keyword' do
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-fetch-KEY: \nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-fetch-KEY: "
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1560,7 +1560,7 @@ describe 'user sends keyword' do
     }
     mail.gpg(gpg_opts)
     content_body = "Hello again!\n"
-    mail.body = "x-list-name: #{list.email}\nX-resend: someone@example.org\nx-stop\n#{content_body}"
+    mail.body = "x-list-name: #{list.email}\nX-resend: someone@example.org\n\n#{content_body}"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1614,7 +1614,7 @@ Otherwise it won't be sent out. What a nice trick!
 
 Best
 EOS
-    mail.body = "x-list-name: #{list.email}\nX-resend: someone@example.org\nx-stop\n#{content_body}"
+    mail.body = "x-list-name: #{list.email}\nX-resend: someone@example.org\n\n#{content_body}"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1640,6 +1640,42 @@ EOS
     teardown_list_and_mailer(list)
   end
 
+  it 'resend returns an error about wrong arguments if email content got into the arguments due to no blank line' do
+    list = create(:list, public_footer: "-- \nblablabla")
+    list.subscribe('schleuder@example.org', '59C71FB38AEE22E091C78259D06350440F759BD3', true)
+    ENV['GNUPGHOME'] = list.listdir
+    mail = Mail.new
+    mail.to = list.email
+    mail.from = list.admins.first.email
+    gpg_opts = {
+      encrypt: true,
+      keys: {list.email => list.fingerprint},
+      sign: true,
+      sign_as: list.admins.first.fingerprint
+    }
+    mail.gpg(gpg_opts)
+    mail.body = "x-list-name: #{list.email}\nX-resend: someone@example.org\nHello Again!\n\nLooking forward to meeting you!"
+    mail.deliver
+
+    encrypted_mail = Mail::TestMailer.deliveries.first
+    Mail::TestMailer.deliveries.clear
+
+    begin
+      Schleuder::Runner.new().run(encrypted_mail.to_s, list.email)
+    rescue SystemExit
+    end
+    expect(Mail::TestMailer.deliveries.length).to eql(1)
+    raw = Mail::TestMailer.deliveries.last
+    message = Mail.create_message_to_list(raw.to_s, list.email, list).setup
+
+    expect(message.to).to eql([list.admins.first.email])
+    expect(message.to_s).to include('Error: Invalid email-address for resending: hello')
+    expect(message.to_s).to include('Error: Invalid email-address for resending: again!')
+    expect(message.to_s).not_to include('Resent: ')
+
+    teardown_list_and_mailer(list)
+  end
+
   it 'x-resend does not include internal_footer' do
     list = create(
       :list,
@@ -1659,7 +1695,7 @@ EOS
     }
     mail.gpg(gpg_opts)
     content_body = "Hello again!\n"
-    mail.body = "x-list-name: #{list.email}\nX-resend: someone@example.org\nx-stop\n#{content_body}"
+    mail.body = "x-list-name: #{list.email}\nX-resend: someone@example.org\n\n#{content_body}"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1695,7 +1731,7 @@ EOS
     mail.gpg(gpg_opts)
     content_body = "Hello again! ¡Hola!\r\n"
     mail.charset = 'iso-8859-1'
-    mail.body = "x-list-name: #{list.email}\nX-resend: someone@example.org\nx-stop\n#{content_body}".encode('iso-8859-1')
+    mail.body = "x-list-name: #{list.email}\nX-resend: someone@example.org\n\n#{content_body}".encode('iso-8859-1')
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1737,7 +1773,7 @@ EOS
     mail.gpg(gpg_opts)
     content_body = "This is a test\r\nAnd here are some umlauts:ÄäÖöÜüß"
     mail.charset = 'utf-8'
-    mail.body = "x-list-name: #{list.email}\nX-resend: someone@example.org\nx-stop\n#{content_body}".encode('utf-8')
+    mail.body = "x-list-name: #{list.email}\nX-resend: someone@example.org\n\n#{content_body}".encode('utf-8')
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1777,7 +1813,7 @@ EOS
     }
     mail.gpg(gpg_opts)
     content_body = "Hello again!\n"
-    mail.body = "x-list-name: #{list.email}\nX-resend: someone@example.org\nx-stop\n#{content_body}"
+    mail.body = "x-list-name: #{list.email}\nX-resend: someone@example.org\n\n#{content_body}"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1819,7 +1855,7 @@ EOS
     }
     mail.gpg(gpg_opts)
     content_body = "Hello again!\n"
-    mail.body = "x-list-name: #{list.email}\nX-resend: someone@example.org\nx-stop\n#{content_body}"
+    mail.body = "x-list-name: #{list.email}\nX-resend: someone@example.org\n\n#{content_body}"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1861,7 +1897,7 @@ EOS
     }
     mail.gpg(gpg_opts)
     content_body = "Hello again!\n"
-    mail.body = "x-list-name: #{list.email}\nX-resend-encrypted-only: bla@foo\n#{content_body}"
+    mail.body = "x-list-name: #{list.email}\nX-resend-encrypted-only: bla@foo\n\n#{content_body}"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1900,7 +1936,7 @@ EOS
     }
     mail.gpg(gpg_opts)
     content_body = "Hello again!\n"
-    mail.body = "x-list-name: #{list.email}\nX-resend-cc-encrypted-only: bla@foo\n#{content_body}"
+    mail.body = "x-list-name: #{list.email}\nX-resend-cc-encrypted-only: bla@foo\n\n#{content_body}"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1939,7 +1975,7 @@ EOS
     }
     mail.gpg(gpg_opts)
     content_body = "Hello again!\n"
-    mail.body = "x-list-name: #{list.email}\nX-resend-cc-encrypted-only: schleuder@example.org bla@foo\n#{content_body}"
+    mail.body = "x-list-name: #{list.email}\nX-resend-cc-encrypted-only: schleuder@example.org bla@foo\n\n#{content_body}"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -1979,7 +2015,7 @@ EOS
     }
     mail.gpg(gpg_opts)
     content_body = "Hello again!\n"
-    mail.body = "x-list-name: #{list.email}\nX-resend-cc-encrypted-only: schleuder@example.org bla@foo\n#{content_body}"
+    mail.body = "x-list-name: #{list.email}\nX-resend-cc-encrypted-only: schleuder@example.org bla@foo\n\n#{content_body}"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -2020,7 +2056,7 @@ Error: Resending to <schleuder@example.org> aborted due to other errors.")
     }
     mail.gpg(gpg_opts)
     content_body = "Hello again!\n"
-    mail.body = "x-list-name: #{list.email}\nX-resend-cc-encrypted-only: schleuder@example.org foo@bla bla@foo\n#{content_body}"
+    mail.body = "x-list-name: #{list.email}\nX-resend-cc-encrypted-only: schleuder@example.org foo@bla bla@foo\n\n\n#{content_body}"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -2057,7 +2093,7 @@ Error: Resending to <bla@foo> aborted due to other errors.")
     }
     mail.gpg(gpg_opts)
     content_body = "Hello again!\n"
-    mail.body = "x-list-name: #{list.email}\nX-resend-cc: foo@bla bla@foo\n#{content_body}"
+    mail.body = "x-list-name: #{list.email}\nX-resend-cc: foo@bla bla@foo\n\n#{content_body}"
     expect(list.keys('foo@bla').size).to eql(0)
     expect(list.keys('bla@foo').size).to eql(0)
     mail.deliver
@@ -2096,7 +2132,7 @@ Error: Resending to <bla@foo> aborted due to other errors.")
     }
     mail.gpg(gpg_opts)
     content_body = "Hello again!\n"
-    mail.body = "x-list-name: #{list.email}\nX-resend-cc: foo@bla bla@foo\n#{content_body}"
+    mail.body = "x-list-name: #{list.email}\nX-resend-cc: foo@bla bla@foo\n\n#{content_body}"
     expect(list.keys('foo@bla').size).to eql(0)
     expect(list.keys('bla@foo').size).to eql(1)
     mail.deliver
@@ -2134,7 +2170,7 @@ Error: Resending to <bla@foo> aborted due to other errors.")
     }
     mail.gpg(gpg_opts)
     content_body = "Hello again!\n"
-    mail.body = "x-list-name: #{list.email}\nX-resend-cc-encrypted-only: foo@bla bla@foo\n#{content_body}"
+    mail.body = "x-list-name: #{list.email}\nX-resend-cc-encrypted-only: foo@bla bla@foo\n\n#{content_body}"
     expect(list.keys('foo@bla').size).to eql(0)
     expect(list.keys('bla@foo').size).to eql(0)
     mail.deliver
@@ -2176,7 +2212,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
     }
     mail.gpg(gpg_opts)
     content_body = "Hello again!\n"
-    mail.body = "x-list-name: #{list.email}\nX-resend-encrypted-only: bla@foo\nx-stop\n#{content_body}"
+    mail.body = "x-list-name: #{list.email}\nX-resend-encrypted-only: bla@foo\n\n#{content_body}"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -2215,7 +2251,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
     }
     mail.gpg(gpg_opts)
     content_body = "Hello again!\n"
-    mail.body = "x-list-name: #{list.email}\nX-resend-unencrypted: bla@foo\nx-stop\n#{content_body}"
+    mail.body = "x-list-name: #{list.email}\nX-resend-unencrypted: bla@foo\n\n#{content_body}"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -2254,7 +2290,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
     }
     mail.gpg(gpg_opts)
     content_body = "Hello again!\n"
-    mail.body = "x-list-name: #{list.email}\nX-resend-encrypted-only: bla@foo\nx-stop\n#{content_body}"
+    mail.body = "x-list-name: #{list.email}\nX-resend-encrypted-only: bla@foo\n\n#{content_body}"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -2289,7 +2325,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
     }
     mail.gpg(gpg_opts)
     content_body = "Hello again!\n"
-    mail.body = "x-list-name: #{list.email}\nX-resend-cc-encrypted-only: bla@foo\n#{content_body}"
+    mail.body = "x-list-name: #{list.email}\nX-resend-cc-encrypted-only: bla@foo\n\n#{content_body}"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -2324,7 +2360,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
     mail.gpg(gpg_opts)
     content_body = "Hello again!\n"
     invalid_recipient = '`ls`bla'
-    mail.body = "x-list-name: #{list.email}\nX-resend: #{invalid_recipient}\nx-stop\n#{content_body}"
+    mail.body = "x-list-name: #{list.email}\nX-resend: #{invalid_recipient}\n\n#{content_body}"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -2360,7 +2396,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
     }
     mail.gpg(gpg_opts)
     signed_text = "signed\nsigned\nsigned\n\n"
-    mail.body = "x-list-name: #{list.email}\nx-sign-this:\nx-stop\n#{signed_text}"
+    mail.body = "x-list-name: #{list.email}\nx-sign-this:\n\n#{signed_text}"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -2393,7 +2429,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
     }
     mail.gpg(gpg_opts)
     keywords = Mail::Part.new
-    keywords.body = "\n\nx-list-name: #{list.email}\nx-sign-this:\nx-stop"
+    keywords.body = "\n\nx-list-name: #{list.email}\nx-sign-this:"
     mail.parts << keywords
     signed_content = File.read('spec/fixtures/example_key.txt')
     mail.attachments['example_key.txt'] = { mime_type: 'application/pgp-key',
@@ -2440,7 +2476,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-list-KEYs: der@ex\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-list-KEYs: der@ex"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -2474,7 +2510,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-list-KEYs: @schleuder\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-list-KEYs: @schleuder"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -2508,7 +2544,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-list-KEYs: 0x59C71FB38AEE22E091C78259D06350440F759BD3\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-list-KEYs: 0x59C71FB38AEE22E091C78259D06350440F759BD3"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -2543,7 +2579,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-GET-KEY: 0x59C71FB38AEE22E091C78259D06350440F759BD3\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-GET-KEY: 0x59C71FB38AEE22E091C78259D06350440F759BD3"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -2577,7 +2613,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-get-KEY: blabla\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-get-KEY: blabla"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -2611,7 +2647,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-get-KEY:\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-get-KEY:"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -2646,7 +2682,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-get-logfile\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-get-logfile"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -2682,7 +2718,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-get-logfile\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-get-logfile"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -2717,7 +2753,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
     }
     mail.gpg(gpg_opts)
     content_body = 'something something list-key'
-    mail.body = "x-list-name: #{list.email}\nX-attach-listkey\nx-stop\n#{content_body}"
+    mail.body = "x-list-name: #{list.email}\nX-attach-listkey\n\n#{content_body}"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -2754,6 +2790,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
     raw = Mail::TestMailer.deliveries.first
     message = Mail.create_message_to_list(raw.to_s, list.email, list).setup
 
+    expect(res).to be_nil
     expect(message.parts.length).to eql(2)
     expect(message.parts.last.parts.length).to eql(2)
     expect(message.parts.last.parts.first.decoded).to eql("Hallo\r\n\r\nkurz mal testen, wie ein resend mail, wo zusätzlich der listkey attached\r\nist bei euch so ankommt.\r\n\r\nich habe das gefühl hier ist as broken.\r\n\r\n\r\n\r\n\r\n")
@@ -2779,7 +2816,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-get-version\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-get-version"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -2813,7 +2850,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-get-version\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-get-version"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -2849,7 +2886,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-list-keys\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-list-keys"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -2886,7 +2923,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-list-keys schleuder2\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-list-keys schleuder2"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -2923,7 +2960,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
       sign_as: list.admins.first.fingerprint
     }
     mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-list-keys schleuder2 bla\nx-stop"
+    mail.body = "x-list-name: #{list.email}\nX-list-keys schleuder2 bla"
     mail.deliver
 
     encrypted_mail = Mail::TestMailer.deliveries.first
@@ -2961,7 +2998,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
         sign_as: list.admins.first.fingerprint
       }
       mail.gpg(gpg_opts)
-      mail.body = "x-list-name: #{list.email}\nX-list-keys\nx-stop"
+      mail.body = "x-list-name: #{list.email}\nX-list-keys"
       mail.deliver
 
       encrypted_mail = Mail::TestMailer.deliveries.first
@@ -2996,7 +3033,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
       }
       mail.gpg(gpg_opts)
       keymaterial = File.read('spec/fixtures/broken_utf8_uid_key.txt')
-      mail.body = "x-list-name: #{list.email}\nX-ADD-KEY:\nx-stop\n#{keymaterial}"
+      mail.body = "x-list-name: #{list.email}\nX-ADD-KEY:\n\n#{keymaterial}"
       mail.deliver
 
       encrypted_mail = Mail::TestMailer.deliveries.first
@@ -3030,7 +3067,7 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
         sign_as: list.admins.first.fingerprint
       }
       mail.gpg(gpg_opts)
-      mail.body = "x-list-name: #{list.email}\nX-GET-KEY: 0x3102B29989BEE703AE5ED62E1242F6E13D8EBE4A\nx-stop"
+      mail.body = "x-list-name: #{list.email}\nX-GET-KEY: 0x3102B29989BEE703AE5ED62E1242F6E13D8EBE4A"
       mail.deliver
 
       encrypted_mail = Mail::TestMailer.deliveries.first
@@ -3048,39 +3085,6 @@ Error: Resending to <bla@foo> failed (0 keys found, of which 0 can be used.
 
       teardown_list_and_mailer(list)
     end
-  end
-
-  it 'rejects messages to request-address without x-stop keyword' do
-    list = create(:list)
-    list.subscribe('schleuder@example.org', '59C71FB38AEE22E091C78259D06350440F759BD3', true)
-    ENV['GNUPGHOME'] = list.listdir
-    mail = Mail.new
-    mail.to = list.request_address
-    mail.from = list.admins.first.email
-    gpg_opts = {
-      encrypt: true,
-      keys: {list.request_address => list.fingerprint},
-      sign: true,
-      sign_as: list.admins.first.fingerprint
-    }
-    mail.gpg(gpg_opts)
-    mail.body = "x-list-name: #{list.email}\nX-LIST-KEYS"
-    mail.deliver
-
-    encrypted_mail = Mail::TestMailer.deliveries.first
-    Mail::TestMailer.deliveries.clear
-
-    begin
-      Schleuder::Runner.new().run(encrypted_mail.to_s, list.request_address)
-    rescue SystemExit
-    end
-    raw = Mail::TestMailer.deliveries.first
-    message = Mail.create_message_to_list(raw.to_s, list.request_address, list).setup
-
-    expect(Mail::TestMailer.deliveries.size).to eql(1)
-    expect(message.first_plaintext_part.body.to_s).to include("Your message lacked the keyword 'X-STOP'")
-
-    teardown_list_and_mailer(list)
   end
 
 end